refactored configuration to be cleaner
