+++ agentops/__init__.py
@@ -1,5 +1,4 @@
 from .client import Client
 from .event import Event, EventState
 from .session import SessionState
-from .config import Configuration
 from .logger import AgentOpsLogger
