+++ agentops/session.py
@@ -1,4 +1,5 @@
 from .helpers import get_ISO_time
+from typing import Optional, List
 
 
 class SessionState:
