+++ tests/test_canary.py
@@ -2,7 +2,7 @@ import pytest
 import requests_mock
 import time
 
-from agentops import Client, Event, Configuration, SessionState
+from agentops import Client, Event, SessionState
 
 
 @pytest.fixture
@@ -18,8 +18,7 @@ class TestCanary:
     def setup_method(self):
         self.url = 'https://agentops-server-v2.fly.dev'
         self.api_key = "random_api_key"
-        self.config = Configuration(max_wait_time=5)
-        self.client = Client(api_key=self.api_key, config=self.config)
+        self.client = Client(api_key=self.api_key, max_wait_time=5)
 
     def teardown_method(self):
         self.client.end_session(end_state=SessionState.SUCCESS)
