+++ tests/test_logger.py
@@ -1,8 +1,7 @@
 import pytest
 import requests_mock
 import time
-from agentops import Client, AgentOpsLogger, Configuration, SessionState
-from agentops.event import SessionState
+from agentops import Client, AgentOpsLogger, SessionState
 
 
 @pytest.fixture
@@ -16,10 +15,9 @@ def mock_req():
 
 class TestLogger:
     def setup_method(self):
-        self.url = 'https://agentops-server-v2.fly.dev'`
+        self.url = 'https://agentops-server-v2.fly.dev'
         self.api_key = "random_api_key"
-        self.config = Configuration(max_wait_time=5)
-        self.client = Client(api_key=self.api_key, config=self.config)
+        self.client = Client(api_key=self.api_key, max_wait_time=5)
 
     def teardown_method(self):
         self.client.end_session(end_state=SessionState.SUCCESS)
